requirements.txt
line-bot-sdk==3.11.0
anthropic==0.25.0
flask==3.0.3
